+++ module/User/config/rbac.config.php
@@ -7,13 +7,13 @@ return [
     'zfc_rbac' => [
         'guards' => [
             RoutePermissionsGuard::class => [
-                'signin' => ['user/auth/signin'],
-                'logout' => ['user/auth/logout'],
-                'signup' => ['user/signup/signup'],
-                'confirm-email' => ['user/confirm-email/confirm'],
-                'confirm-again' => ['user/confirm-again/again'],
-                'forgot-pass' => ['user/access/forgot-pass'],
-                'restore-pass' => ['user/access/restore-pass'],
+                'signin' => 'user/auth/signin',
+                'logout' => 'user/auth/logout',
+                'signup' => 'user/signup/signup',
+                'confirm-email' => 'user/confirm-email/confirm',
+                'confirm-again' => 'user/confirm-again/again',
+                'forgot-pass' => 'user/access/forgot-pass',
+                'restore-pass' => 'user/access/restore-pass',
             ],
         ],
         'role_provider' => [