+++ module/User/test/_fixture/UserFixture.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace User\Test;
+
+use ExCodeception\BaseFixture;
+
+class UserFixture extends BaseFixture
+{
+    protected $dataFile = 'user';
+    protected $entityClass = 'User\Entity\User';
+}